+++ ReleaseNotes/release_notes.md
@@ -7,6 +7,8 @@
 ## Backward incompatible:  
 
 ## Bug fixes:
+ 1. fix formfactor for `Ostap::Math::ChannelFlatteBugg`
+
 
 # v1.9.3.8